[core] Minor refactoring to events and structure
