+++ src/main/java/io/javalin/http/CachedRequestWrapper.kt
@@ -4,7 +4,7 @@
  * Licensed under Apache 2.0: https://github.com/tipsy/javalin/blob/master/LICENSE
  */
 
-package io.javalin.core
+package io.javalin.http
 
 import io.javalin.Javalin
 import java.io.ByteArrayInputStream
